% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psych_fxns.R
\name{desc_data}
\alias{desc_data}
\title{This function describe data using psych's basic functions}
\usage{
desc_data(dataset)
}
\arguments{
\item{dataset}{A dataset}
}
\value{
result table of described dataset
}
\description{
This function describe data using psych's basic functions
}
