% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psych_fxns.R
\name{draw_Corr}
\alias{draw_Corr}
\title{This function uses:
- paris.panels()
- corPlot()
to graph the correlations to scatter plot and heatmap}
\usage{
draw_Corr(dataset.num.col, lowerCor)
}
\value{
It will create 2 png files of each
}
\description{
This function uses:
- paris.panels()
- corPlot()
to graph the correlations to scatter plot and heatmap
}
\examples{
draw_Corr(num_col_dataset, lowerCor_var)
}
