% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psych_fxns.R
\name{get_num_col}
\alias{get_num_col}
\title{This function retrives and peeks numeric columns data}
\usage{
get_num_col(dataset)
}
\arguments{
\item{dataset}{A dataset}
}
\value{
dataset with numerical columns
}
\description{
This function retrives and peeks numeric columns data
}
