% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_fxns.R
\name{get_number_of_songBy_yr}
\alias{get_number_of_songBy_yr}
\title{This function returns number of songs grouped by year}
\usage{
get_number_of_songBy_yr(dataset)
}
\arguments{
\item{dataset}{The dataset}
}
\value{
number of songs by year
}
\description{
This function returns number of songs grouped by year
}
\examples{
get_number_of_songBy_yr(songs)
}
