% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_fxns.R
\name{get_popSongs}
\alias{get_popSongs}
\title{This function returns the most popular songs over the decade}
\usage{
get_popSongs(dataset)
}
\arguments{
\item{dataset}{The dataset}
}
\value{
songs with pop value over 80
}
\description{
This function returns the most popular songs over the decade
}
\examples{
get_popSongs(songs)
}
