% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_fxns.R
\name{get_topArtists}
\alias{get_topArtists}
\title{This function returns all top 10 artists of past 10 years}
\usage{
get_topArtists(dataset)
}
\arguments{
\item{dataset}{The dataset}
}
\value{
conbined results of top_artist_yr()
}
\description{
This function returns all top 10 artists of past 10 years
}
\examples{
top_artist_yr_all(dataset)
}
