% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_fxns.R
\name{get_topArtistsBy_yr}
\alias{get_topArtistsBy_yr}
\title{######################################################
POP value is the popularity of the artist.
The value will be between 0 and 100, with 100 being the most popular.
The artist’s popularity is calculated from the popularity of all the artist’s tracks.
With no steaming data available in this dataset.
The artist's popularity are deternmined by the weighted value:}
\usage{
get_topArtistsBy_yr(dataset, y)
}
\arguments{
\item{dataset}{The dataset}

\item{y}{The year}
}
\description{
1. POP value of the artist at that time * (20%)
2. The number of songs of the artist appeared on the list of that year (80%)
#####################################################
}
\details{
This function returns top artist of the year
}
\examples{
top_artist_yr(dataset, 2012)
}
