% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_fxns.R
\name{get_topDnce}
\alias{get_topDnce}
\title{This function returns top dance songs
evaluated by (dnce by .3 + pop by .4  + nrgy by .3)}
\usage{
get_topDnce(dataset)
}
\arguments{
\item{dataset}{The dataset}
}
\description{
The dancibility are re-evaluated on top of original
dnce value with weight given to {nrgy} and {bpm}
}
\examples{
top_dnce_by_year(songs, 2012)
}
