% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psych_fxns.R
\name{lowerCorr}
\alias{lowerCorr}
\title{This function prints corr test result using corr.test() to
Find and print (usinglowerMat) the lower diagonal correlation matrix
but returns (in-visibly) the full correlation matrix found with the use}
\usage{
lowerCorr(dataset)
}
\arguments{
\item{dataset}{A dataset}
}
\value{
results table of corr.test()
}
\description{
This function prints corr test result using corr.test() to
Find and print (usinglowerMat) the lower diagonal correlation matrix
but returns (in-visibly) the full correlation matrix found with the use
}
