% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psych_fxns.R
\name{testCorr}
\alias{testCorr}
\title{This function prints corr test result using corr.test() to
Find the correlations, sample sizes,  and probability values
between elements of a matrix or data.frame.}
\usage{
testCorr(dataset.num.col)
}
\arguments{
\item{dataset}{A dataset}
}
\value{
results table of corr.test()
}
\description{
This function prints corr test result using corr.test() to
Find the correlations, sample sizes,  and probability values
between elements of a matrix or data.frame.
}
